rename languageService folder to languageServiceWrapper to avoid confusions
