+++ package/src/languageServiceWrapper/kustoLanguageService.ts
