+++ package/src/languageServiceWrapper/kustoMonarchLanguageDefinition.ts
