+++ package/src/languageServiceWrapper/settings.ts
