+++ package/src/monaco.contribution.ts
@@ -11,12 +11,12 @@ import {
     getExpression,
     getInputParametersAsCslString,
     getEntityDataTypeFromCslType,
-} from './languageService/schema';
-import type { LanguageSettings } from './languageService/settings';
+} from './languageServiceWrapper/schema';
+import type { LanguageSettings } from './languageServiceWrapper/settings';
 
-export * from './languageService/schema';
-export * from './languageService/renderInfo';
-export * from './languageService/settings';
+export * from './languageServiceWrapper/schema';
+export * from './languageServiceWrapper/renderInfo';
+export * from './languageServiceWrapper/settings';
 export * from './types';
 
 export { getCurrentCommandRange } from './extendedEditor';
